Created a stub method in UserController for approving volunteers
Will call JdbcVolunteerDao and update their approval_status to 1
and will call JdbcUserDao to create a new user with some default password
and will return that user to the frontend

I think returning a user is the appropriate way to go, so that the end-user
can see what their default password is or something.

Or maybe not, since the actual end-user isn't the one approving their own
account, it may be better for it to be a null method.